Split backend liveness test in three files, fix missing d3d12 flush on resize
